Moving to height_energy
