+++ aviary/interface/test/test_reserve_support.py
@@ -5,7 +5,7 @@ from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
 
 from aviary.interface.methods_for_level2 import AviaryProblem
-from aviary.interface.default_phase_info.simple import phase_info as ph_in_flops
+from aviary.interface.default_phase_info.height_energy import phase_info as ph_in_flops
 from aviary.interface.default_phase_info.two_dof import phase_info as ph_in_gasp
 from aviary.variable_info.variables import Aircraft, Mission
