+++ aviary/interface/utils/check_phase_info.py
@@ -2,7 +2,7 @@ from openmdao.utils.units import valid_units
 from aviary.variable_info.enums import SpeedType, EquationsOfMotion
 
 TWO_DEGREES_OF_FREEDOM = EquationsOfMotion.TWO_DEGREES_OF_FREEDOM
-SIMPLE = EquationsOfMotion.SIMPLE
+HEIGHT_ENERGY = EquationsOfMotion.HEIGHT_ENERGY
 SOLVED = EquationsOfMotion.SOLVED
 
 
@@ -193,7 +193,7 @@ def check_phase_info(phase_info, mission_method):
                 phase_keys[phase] = phase_keys_gasp[phase]
     elif mission_method is SOLVED:
         return
-    elif mission_method is SIMPLE:
+    elif mission_method is HEIGHT_ENERGY:
         for phase in phase_info:
             if phase != 'pre_mission' and phase != 'post_mission':
                 phase_keys[phase] = {**common_keys, **common_entries}
