+++ aviary/mission/flops_based/ode/test/test_takeoff_ode.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 
 import openmdao.api as om
 
-from aviary.interface.default_phase_info.simple import default_mission_subsystems
+from aviary.interface.default_phase_info.height_energy import default_mission_subsystems
 from aviary.mission.flops_based.ode.takeoff_ode import TakeoffODE
 from aviary.models.N3CC.N3CC_data import (
     detailed_takeoff_climbing, detailed_takeoff_ground, takeoff_subsystem_options, inputs)
