+++ aviary/models/N3CC/N3CC_data.py
@@ -26,7 +26,7 @@ from aviary.variable_info.variables import Dynamic as _Dynamic
 from aviary.variable_info.variables import Mission
 from aviary.variable_info.variables import Settings
 from aviary.variable_info.enums import EquationsOfMotion, LegacyCode
-from aviary.interface.default_phase_info.simple import default_mission_subsystems
+from aviary.interface.default_phase_info.height_energy import default_mission_subsystems
 
 Dynamic = _Dynamic.Mission
 
@@ -315,7 +315,7 @@ inputs.set_val(Mission.Takeoff.THRUST_INCIDENCE, 0., 'deg')
 
 # Settings
 # ---------------------------
-inputs.set_val(Settings.EQUATIONS_OF_MOTION, EquationsOfMotion.SIMPLE)
+inputs.set_val(Settings.EQUATIONS_OF_MOTION, EquationsOfMotion.HEIGHT_ENERGY)
 inputs.set_val(Settings.MASS_METHOD, LegacyCode.FLOPS)
 
 # ---------------------------
