+++ aviary/models/N3CC/N3CC_generic_low_speed_polars_FLOPSinp.csv
@@ -169,7 +169,7 @@ mission:takeoff:rolling_friction_coefficient,0.0175,unitless
 mission:takeoff:spoiler_drag_coefficient,0.085,unitless
 mission:takeoff:spoiler_lift_coefficient,-0.81,unitless
 mission:takeoff:thrust_incidence,0,deg
-settings:equations_of_motion,simple
+settings:equations_of_motion,height_energy
 settings:mass_method,FLOPS
 
 # Unconverted Values
