+++ aviary/models/test_aircraft/aircraft_for_bench_FwFm_simple.csv
@@ -155,5 +155,5 @@ mission:takeoff:fuel_simple,577,lbm
 mission:takeoff:lift_coefficient_max,3.0,unitless
 mission:takeoff:lift_over_drag,17.354,unitless
 mission:takeoff:rolling_friction_coefficient,0.0175,unitless
-settings:equations_of_motion,simple
+settings:equations_of_motion,height_energy
 settings:mass_method,FLOPS
