+++ aviary/subsystems/test/test_external_subsystem_bus.py
@@ -9,7 +9,7 @@ import numpy as np
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 
-from aviary.interface.default_phase_info.simple import phase_info as ph_in
+from aviary.interface.default_phase_info.height_energy import phase_info as ph_in
 from aviary.interface.methods_for_level2 import AviaryProblem
 from aviary.subsystems.subsystem_builder_base import SubsystemBuilderBase
