+++ aviary/validation_cases/benchmark_tests/test_0_iters.py
@@ -5,7 +5,7 @@ from openmdao.utils.testing_utils import require_pyoptsparse, use_tempdirs
 
 from aviary.interface.methods_for_level2 import AviaryProblem
 from aviary.interface.default_phase_info.two_dof import phase_info as two_dof_phase_info
-from aviary.interface.default_phase_info.simple import phase_info as simple_phase_info
+from aviary.interface.default_phase_info.height_energy import phase_info as simple_phase_info
 from aviary.interface.default_phase_info.solved import phase_info as solved_phase_info
 from aviary.variable_info.enums import EquationsOfMotion
