+++ aviary/validation_cases/benchmark_tests/test_FLOPS_based_sizing_N3CC.py
@@ -22,11 +22,11 @@ from aviary.validation_cases.validation_tests import get_flops_inputs
 from aviary.variable_info.functions import setup_trajectory_params
 from aviary.utils.aviary_values import AviaryValues
 from aviary.variable_info.variables_in import VariablesIn
-from aviary.mission.flops_based.phases.simple_energy_phase import EnergyPhase
+from aviary.mission.flops_based.phases.energy_phase import EnergyPhase
 
 from aviary.variable_info.variables import Aircraft, Dynamic, Mission
 from aviary.subsystems.premission import CorePreMission
-from aviary.interface.default_phase_info.simple import default_premission_subsystems, default_mission_subsystems
+from aviary.interface.default_phase_info.height_energy import default_premission_subsystems, default_mission_subsystems
 from aviary.utils.preprocessors import preprocess_crewpayload
 
 try:
