+++ aviary/validation_cases/benchmark_tests/test_FLOPS_detailed_landing.py
@@ -18,7 +18,7 @@ from aviary.models.N3CC.N3CC_data import (
     landing_fullstop_user_options as _landing_fullstop_user_options)
 
 from aviary.variable_info.variables import Dynamic as _Dynamic
-from aviary.interface.default_phase_info.simple import default_premission_subsystems
+from aviary.interface.default_phase_info.height_energy import default_premission_subsystems
 from aviary.utils.preprocessors import preprocess_crewpayload
 from aviary.variable_info.variables_in import VariablesIn
