+++ aviary/validation_cases/benchmark_tests/test_bench_GwFm.py
@@ -15,7 +15,7 @@ from aviary.interface.methods_for_level1 import run_aviary
 from aviary.validation_cases.benchmark_utils import \
     compare_against_expected_values
 from aviary.variable_info.variables import Dynamic
-from aviary.interface.default_phase_info.simple import phase_info
+from aviary.interface.default_phase_info.height_energy import phase_info
 
 
 @use_tempdirs
